yash: parser: 複合コマンドの内容は空でもよいように yash: それに伴うテストの変更・追加, +tests/parser.y
git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@885 048f04df-13f5-43d7-8114-9f9ceecaec24

+++ tests/alias.y.tst
@@ -23,6 +23,18 @@ alias foobar='FOO=BAR ' e='env'
 alias -g G='|grep'
 foobar e G '^FOO='
 
+unalias -a
+alias -g N='>/dev/null'
+alias -p \N
+echo null N  # prints nothing
+N echo null  # prints nothing
+(echo null) N  # prints nothing
+{ echo null; } N  # prints nothing
+if :; then echo null; fi N  # prints nothing
+if if :; then :; fi then if :; then echo null; fi fi N  # prints nothing
+if if :; then :; fi then if :; then echo null; fi N fi  # prints nothing
+if if echo null; then :; fi N then if :; then :; fi fi  # prints nothing
+
 command -V alias unalias
 
 rm -f "$tmp"
